+++ zetaclient/config/config.go
@@ -9,13 +9,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/asaskevich/govalidator"
-	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/fsnotify/fsnotify"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
-
-	"github.com/zeta-chain/node/pkg/chains"
 )
 
 // restrictedAddressBook is a map of restricted addresses
@@ -33,7 +29,7 @@ const folder string = "config"
 // Save saves ZetaClient config
 func Save(config *Config, path string) error {
 	// validate config
-	if err := Validate(*config); err != nil {
+	if err := config.Validate(); err != nil {
 		return errors.Wrapf(err, "config file validation failed")
 	}
 
@@ -91,68 +87,11 @@ func Load(basePath string) (Config, error) {
 	cfg.ZetaCoreHome = basePath
 
 	// validate config
-	if err := Validate(cfg); err != nil {
+	if err := cfg.Validate(); err != nil {
 		return Config{}, errors.Wrapf(err, "config file validation failed")
 	}
 
 	return cfg, nil
-}
-
-// Validate performs basic validation on the config fields
-func Validate(cfg Config) error {
-	// go-tss requires a valid IPv4 address
-	if cfg.PublicIP != "" && !govalidator.IsIPv4(cfg.PublicIP) {
-		return errors.Errorf("reason: invalid public IP, got: %s", cfg.PublicIP)
-	}
-
-	if cfg.PublicDNS != "" && !govalidator.IsDNSName(cfg.PublicDNS) {
-		return errors.Errorf("reason: invalid public DNS, got: %s", cfg.PublicDNS)
-	}
-
-	if _, err := chains.ZetaChainFromCosmosChainID(cfg.ChainID); err != nil {
-		return errors.Errorf("reason: invalid chain id, got: %s", cfg.ChainID)
-	}
-
-	// ZetaCoreURL can be either an IP address or a hostname (e.g., Docker service name)
-	if cfg.ZetaCoreURL != "" && !govalidator.IsIP(cfg.ZetaCoreURL) && !govalidator.IsDNSName(cfg.ZetaCoreURL) {
-		return errors.Errorf("reason: invalid zetacore URL, got: %s", cfg.ZetaCoreURL)
-	}
-
-	// validate granter address - should be a valid bech32 address
-	if _, err := sdktypes.AccAddressFromBech32(cfg.AuthzGranter); err != nil {
-		return errors.Errorf("reason: invalid bech32 granter address, got: %s", cfg.AuthzGranter)
-	}
-
-	// validate grantee name - should not be empty
-	if strings.TrimSpace(cfg.AuthzHotkey) == "" {
-		return errors.Errorf("reason: grantee name is empty")
-	}
-
-	// acceptable log levels are: 0:debug, 1:info, 2:warn, 3:error, 4:fatal, 5:panic
-	if cfg.LogLevel < 0 || cfg.LogLevel > 5 {
-		return errors.Errorf("reason: log level must be between 0 and 5, got: %d", cfg.LogLevel)
-	}
-
-	if cfg.ConfigUpdateTicker == 0 {
-		return errors.Errorf("reason: config update ticker is 0")
-	}
-
-	if cfg.KeyringBackend != KeyringBackendFile && cfg.KeyringBackend != KeyringBackendTest {
-		return errors.Errorf("reason: invalid keyring backend, got: %s", cfg.KeyringBackend)
-	}
-
-	if cfg.MaxBaseFee < 0 {
-		return errors.Errorf("reason: max base fee cannot be negative, got: %d", cfg.MaxBaseFee)
-	}
-
-	if cfg.MempoolCongestionThreshold < 0 {
-		return errors.Errorf(
-			"reason: mempool congestion threshold cannot be negative, got: %d",
-			cfg.MempoolCongestionThreshold,
-		)
-	}
-
-	return nil
 }
 
 // SetRestrictedAddressesFromConfig loads compliance data (restricted addresses) from config.